My proposal for the ReplacingFilterVisitor (now called NormalizeFilters)
